+++ noodles-vcf/src/header/parser/record/value/map.rs
@@ -13,8 +13,10 @@ pub use self::{
     alternative_allele::parse_alternative_allele, contig::parse_contig, filter::parse_filter,
     format::parse_format, info::parse_info, other::parse_other,
 };
+use crate::header::FileFormat;
 
 const PREFIX: u8 = b'<';
+const VCF_4_3: FileFormat = FileFormat::new(4, 3);
 
 /// An error returned when a VCF header record map value fails to parse.
 #[derive(Clone, Debug, Eq, PartialEq)]
@@ -36,8 +38,20 @@ impl fmt::Display for ParseError {
     }
 }
 
-pub fn is_map(src: &[u8]) -> bool {
-    src.first().map(|&b| b == PREFIX).unwrap_or_default()
+pub fn is_map(src: &[u8], file_format: FileFormat) -> bool {
+    const ID: &[u8] = b"ID=";
+
+    fn contains(buf: &[u8], query: &[u8]) -> bool {
+        buf.windows(query.len()).any(|window| window == query)
+    }
+
+    let has_prefix = src.first().is_some_and(|&b| b == PREFIX);
+
+    if file_format < VCF_4_3 {
+        has_prefix && contains(src, ID)
+    } else {
+        has_prefix
+    }
 }
 
 pub fn consume_prefix(src: &mut &[u8]) -> Result<(), ParseError> {
@@ -74,8 +88,15 @@ mod tests {
 
     #[test]
     fn test_is_map() {
-        assert!(is_map(b"<ID=noodles>"));
-        assert!(!is_map(b"noodles"));
-        assert!(!is_map(b""));
+        const VCF_4_2: FileFormat = FileFormat::new(4, 2);
+
+        assert!(is_map(b"<ID=noodles>", VCF_4_2));
+        assert!(!is_map(b"<noodles>", VCF_4_2));
+        assert!(!is_map(b"noodles", VCF_4_2));
+        assert!(!is_map(b"", VCF_4_2));
+
+        assert!(is_map(b"<ID=noodles>", VCF_4_3));
+        assert!(!is_map(b"noodles", VCF_4_3));
+        assert!(!is_map(b"", VCF_4_3));
     }
 }